Several changes to reorganize the code for clarity, as follows:
(1) For clarity, rename ush/source_funcs.sh to ush/source_util_funcs.sh.
(2) Assume that the utility functions that ush/source_util_funcs.sh sources will be moved to a subdirectory named bash_utils under the ush directory.  Note that the actual moving of these files will be done in a later commit.
(3) In ush/source_util_funcs.sh, instead of sourcing the utility functions directly, define a new function named source_util_funcs() that sources these utility functions and then call this function right after each definition.  This allows local convenience variables to be declared in the scope of this new function (such as the name of the file, the directory in which the file is located, etc) that have the same names as variables in the script or function that sources ush/source_util_funcs.sh but that have a different scope so that they don't overwrite the values of the identically-named variables in the sourcing script/function.  This allows error and informational messages to use correct names for files, directories, functions, etc.