Change repo of NEMSfv3gfs code to the NCAR fork of the ufs-weather-model on github, and change the branch to dtc/develop.
In this latest code, the ccpp directory is under the FV3 directory, so make that change where needed.  Also, in this code, there are more stringent checks to ensure the correct ozone production/loss file is being used.  For this purpose, ensure that the newer (after 2015) file is being used with the GFS physics suite and the older (2015) file is being used with the GSD physics suite, and change the template FV3 namelist file for GFS physics so that oz_phys is set to .true. and oz_phys_2015 is set to .false..  Finally, note that in ensuring that the correct ozone production/loss file is being used, it is convenient to have a function that returns to stdout the indices of those elements of a specified array that are equal to a specified string.  Create such a function in the file ush/bash_utils/get_elem_inds.sh, source it in ush/source_util_funcs.sh, and use it in generate_FV3SAR_wflow.sh.