Add capability for tasks to fall back to default modulefile (#204)
This PR adds the capability for tasks to fall back to default modulefile if task-specific modulefile does not exist. In doing so, it gets rid of the old, hacky way I got post to build on Cheyenne, and replaces it with a build and run process that utilizes a new cheyenne.default modulefile.

This should allow us to condense down builds on new machines to use fewer task-specific modulefiles.