Bug fix -- Remove lines at start of rocoto XML.  These cause the rocoto commands to fail. (#247)
DESCRIPTION OF CHANGES:
Blank lines at the start of the rocoto XML apparently cause the rocoto commands to fail. So, in this PR, we modify the rocoto XML template as follows:

Remove all blank lines at start.
Put in a dash with jinja end-of-comment, i.e. change #} to -#}, for the jinja comment at the start of the template so that jinja doesn't leave an extra blank line after processing.

TESTS CONDUCTED:
Conducted one WE2E test on Hera. Rocoto fails to start workflow without these changes, succeeds with.