new weather model executable name (#374)
## DESCRIPTION OF CHANGES: 
The develop branch of ufs-weather-model was updated a couple months ago so that the weather model is no longer built as "NEMS.exe" (https://github.com/ufs-community/ufs-weather-model/pull/217). The workflow generation script needs to be updated with the new model executable name: ufs_model. 

This PR will need to be merged in order to use regional_workflow and the ufs-srweather-app with the most up-to-date develop branch of ufs-weather-model. A PR will be upcoming to update ufs-srweather-app to point to this hash once this PR is merged.

## TESTS CONDUCTED: 
Ran generation script successfully with fully updated workflow.

## ISSUE (optional): 
This is necessary (but not sufficient) in order to resolve https://github.com/ufs-community/ufs-srweather-app/issues/57